ref(translation): Extract translation logic into dedicated app
Relocate TranslationTask model, services, agents, and mixins from the core
app to a new translation app to improve modularity and separation of concerns.

- Move GPTTranslationAgent, TranslationService, and protocols to translation/
- Relocate TranslationTask model and associated migrations
- Move AutomatedTranslationMixin and status mixins to translation/mixins.py
- Consolidate Celery configuration in core/celery.py
- Refine task retry logic to only handle transient RequestExceptions
- Fix admin/task race condition using update_or_create in mixins and tasks
- Update all project-wide imports and test patches
- Clean up core module by removing obsolete code and debug tasks
