+++ backend/astrophotography/serializers.py
@@ -9,8 +9,8 @@ from django.conf import settings
 from django.urls import reverse
 from django.utils import translation
 
-from core.services import TranslationService
 from core.utils.signing import generate_signed_url_params
+from translation.services import TranslationService
 
 from .models import (
     AstroImage,
