+++ backend/core/celery.py
@@ -19,9 +19,3 @@ app.config_from_object("django.conf:settings", namespace="CELERY")
 
 # Load task modules from all registered Django apps.
 app.autodiscover_tasks()
-
-
-@app.task(bind=True, ignore_result=True)
-def debug_task(self):
-    """Debug task to verify Celery is working."""
-    print(f"Request: {self.request!r}")
