--- backend/core/celery/__init__.py
@@ -1,7 +1,0 @@
-# backend/core/celery/__init__.py
-"""
-Celery initialization module.
-"""
-from .app import app
-
-__all__ = ["app"]
