+++ backend/translation/apps.py
@@ -1,0 +1,8 @@
+# backend/translation/apps.py
+from django.apps import AppConfig
+
+
+class TranslationConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "translation"
+    verbose_name = "Translation"
