+++ backend/translation/tasks.py
@@ -1,11 +1,6 @@
-# backend/core/celery/tasks.py
+# backend/translation/tasks.py
 """
 Celery tasks for asynchronous translation processing.
-
-Key Design: Tasks are THIN WRAPPERS around TranslationService methods.
-- Service handles all translation logic (GPT method selection, idempotency, saving)
-- Tasks add async execution + retry capabilities
-- DRY: Single source of truth in TranslationService
 """
 import logging
 from typing import Any
@@ -17,8 +12,8 @@ from django.apps import apps
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
 
-from core.models import TranslationTask
-from core.services import TranslationService
+from .models import TranslationTask
+from .services import TranslationService
 
 logger = logging.getLogger(__name__)
 
@@ -30,40 +25,18 @@ logger = logging.getLogger(__name__)
     retry_backoff=True,
     retry_backoff_max=600,
     retry_jitter=True,
-    autoretry_for=(RequestException,),
 )
 def translate_instance_task(
     self,
     model_name: str,
-    instance_pk: int,
+    instance_pk: Any,
     language_code: str,
     method_name: str,
     **kwargs: Any,
 ) -> dict[str, Any]:
     """
     Async wrapper for TranslationService methods.
-
-    This task delegates ALL translation logic to the service:
-    - Service checks if translation exists (_has_translation)
-    - Service chooses correct GPT handler (translate, translate_html, translate_place, etc.)
-    - Service saves translations with atomic transactions
-
-    This task just adds:
-    - Asynchronous execution via Celery
-    - Retry on network/GPT failures
-    - Background processing (non-blocking admin saves)
-
-    Args:
-        model_name: Full model path (e.g., "astrophotography.AstroImage")
-        instance_pk: Primary key of the instance
-        language_code: Target language code (e.g., "pl", "es")
-        method_name: TranslationService method name (e.g., "translate_place")
-        **kwargs: Extra context passed to service method (e.g., force=True for tags)
-
-    Raises:
-        Retry: On network errors, rate limits, or temporary GPT failures
     """
-
     logger.info(
         "Starting translation task: %s.%s(pk=%s, lang=%s)",
         model_name,
@@ -73,11 +46,11 @@ def translate_instance_task(
     )
 
     try:
-        # 0. Load instance
+        # Load instance
         Model = apps.get_model(model_name)
         instance = Model.objects.get(pk=instance_pk)
 
-        # 1. Create/Update TranslationTask (PENDING -> RUNNING)
+        # Create/Update TranslationTask (PENDING -> RUNNING)
         content_type = ContentType.objects.get_for_model(Model)
         task_record, created = TranslationTask.objects.update_or_create(
             content_type=content_type,
@@ -86,16 +59,16 @@ def translate_instance_task(
             defaults={
                 "method": method_name,
                 "status": TranslationTask.Status.RUNNING,
-                "task_id": self.request.id or str(instance_pk),  # Fallback for local testing
+                "task_id": self.request.id or str(instance_pk),
                 "error_message": "",
             },
         )
 
-        # 2. Call service method (it handles everything: idempotency, GPT, saving)
+        # Call service method
         method = getattr(TranslationService, method_name)
         result = method(instance, language_code, **kwargs)
 
-        # 3. Update Status -> COMPLETED
+        # Update Status -> COMPLETED
         task_record.status = TranslationTask.Status.COMPLETED
         task_record.save()
 
@@ -117,8 +90,16 @@ def translate_instance_task(
 
     except ObjectDoesNotExist:
         logger.error(f"Instance not found: {model_name}(pk={instance_pk}). Skipping translation.")
-        # 3a. Update Status -> FAILED (if record exists, unlikely safely ignored)
-        # Don't retry if object was deleted
+        # Mark as FAILED if record exists
+        try:
+            Model = apps.get_model(model_name)
+            content_type = ContentType.objects.get_for_model(Model)
+            TranslationTask.objects.filter(
+                content_type=content_type, object_id=str(instance_pk), language=language_code
+            ).update(status=TranslationTask.Status.FAILED, error_message="Instance not found")
+        except Exception:
+            pass
+
         return {
             "model": model_name,
             "pk": instance_pk,
@@ -128,7 +109,13 @@ def translate_instance_task(
             "error": "Instance not found",
         }
 
+    except RequestException as exc:
+        # Retry only on network/API transient errors
+        logger.warning(f"Transient error in translation task: {exc}. Retrying...")
+        raise self.retry(exc=exc)
+
     except Exception as exc:
+        # Log and mark as FAILED for all other errors (permanent/dev errors)
         logger.exception(
             "Translation task failed: %s.%s(pk=%s, lang=%s)",
             model_name,
@@ -137,7 +124,6 @@ def translate_instance_task(
             language_code,
         )
 
-        # 3b. Update Status -> FAILED
         try:
             Model = apps.get_model(model_name)
             content_type = ContentType.objects.get_for_model(Model)
@@ -145,7 +131,14 @@ def translate_instance_task(
                 content_type=content_type, object_id=str(instance_pk), language=language_code
             ).update(status=TranslationTask.Status.FAILED, error_message=str(exc)[:500])
         except Exception:
-            pass  # Don't let logging failure mask actual error
+            pass
 
-        # Retry on any other exception (network errors, GPT failures, etc.)
-        raise self.retry(exc=exc)
+        # Do not retry on unknown exceptions to avoid infinite loops of permanent errors
+        return {
+            "model": model_name,
+            "pk": instance_pk,
+            "language": language_code,
+            "method": method_name,
+            "result": None,
+            "error": str(exc),
+        }
